2nd review

+++ recipes/views.py
@@ -1,25 +1,44 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.core.paginator import Paginator
 from .models import (
-    Recipe, User, IngredientRecipe, Follow, Favorite, 
-    Purchase, Ingredient
+    Recipe, User, IngredientRecipe, Ingredient
 )
+from api.models import Follow, Favorite, Purchase
 from .forms import RecipeForm
 from .utils import save_recipe
 from django.contrib.auth.decorators import login_required
 from reportlab.pdfgen import canvas
 from django.http import HttpResponse
+from django.db.models import Q
+from functools import reduce
+import operator
+
+
+def get_tags(request):
+    tags = []
+    tags_filter = None
+    if request.GET.getlist('tag'):
+        tags = request.GET.getlist('tag')
+    if tags:
+        tags_filter = reduce(
+            operator.or_, (Q(tag__contains=tag) for tag in tags))
+    tags = request.GET.getlist('tag')
+    return tags, tags_filter
 
 
 def index(request):
-    recipe_list = Recipe.objects.all()
+    tags, tags_filter = get_tags(request)
+    if tags_filter:
+        recipe_list = Recipe.objects.filter(tags_filter).all()
+    else:
+        recipe_list = Recipe.objects.all()
     paginator = Paginator(recipe_list, 6)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
     return render(
         request,
         'index.html',
-        {'page': page, 'paginator': paginator}
+        {'page': page, 'paginator': paginator, 'tags': tags}
     )
 
 
@@ -34,10 +53,7 @@ def new_recipe(request):
 
 @login_required
 def recipe_edit(request, username, recipe_id):
-    recipe = get_object_or_404(Recipe, id=recipe_id)
-    relation = IngredientRecipe.objects.filter(
-        recipe=recipe,
-        )
+    recipe = get_object_or_404(Recipe, author__username=username, id=recipe_id)
     form = RecipeForm(
         data=request.POST or None,
         files=request.FILES or None,
@@ -49,42 +65,27 @@ def recipe_edit(request, username, recipe_id):
     return render(
         request,
         'recipe_edit.html',
-        {'form': form, 'relation': relation, 'recipe': recipe}
+        {'form': form, 'recipe': recipe}
     )
 
 
 def recipe_view(request, username, recipe_id):
-    author = get_object_or_404(User, username=username)
     recipe = get_object_or_404(Recipe, author__username=username, id=recipe_id)
-    favorite = Favorite.objects.filter(recipe=recipe)
-    purchase = Purchase.objects.filter(recipe=recipe)
-    relation = IngredientRecipe.objects.filter(
-        recipe=recipe,
-        )
     form = RecipeForm()
-    following = None
-    if request.user.is_authenticated:
-        following = Follow.objects.filter(
-            user=request.user,
-            author=author).exists()
     return render(
         request,
         'single_recipe.html',
-        {
-            'recipe': recipe,
-            'author': recipe.author,
-            'relation': relation,
-            'form': form,
-            'following': following,
-            'favorite': favorite,
-            'purchase': purchase
-            }
-        )
+        {'recipe': recipe, 'form': form}
+    )
 
 
 def profile(request, username):
     author = get_object_or_404(User, username=username)
-    recipe_list = Recipe.objects.filter(author=author)
+    tags, tags_filter = get_tags(request)
+    if tags_filter:
+        recipe_list = Recipe.objects.filter(tags_filter).filter(author=author)
+    else:
+        recipe_list = Recipe.objects.filter(author=author)
     paginator = Paginator(recipe_list, 6)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
@@ -100,6 +101,7 @@ def profile(request, username):
             'paginator': paginator,
             'author': author,
             'following': following,
+            'tags': tags,
             }
         )
 
@@ -107,15 +109,15 @@ def profile(request, username):
 @login_required
 def recipe_delete(request, username, recipe_id):
     author = get_object_or_404(User, username=username)
-    recipe = Recipe.objects.filter(author=author, id=recipe_id)
-    if recipe.exists():
+    recipe = get_object_or_404(Recipe, author=author, id=recipe_id)
+    if recipe.exists() and author == request.user:
         recipe.delete()
     return redirect('profile', username=username)
 
 
 @login_required
-def follow_index(request):
-    authors = User.objects.filter(following__user=request.user)
+def follow_index(request, username):
+    authors = User.objects.filter(following__user__username=username)
     paginator = Paginator(authors, 5)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
@@ -127,33 +129,21 @@ def follow_index(request):
 
 
 @login_required
-def profile_follow(request, username):
-    author = get_object_or_404(User, username=username)
-    if request.user != author:
-        Follow.objects.get_or_create(user=request.user, author=author)
-    return redirect('profile', username=username)
-
-
-@login_required
-def profile_unfollow(request, username):
-    author = get_object_or_404(User, username=username)
-    relation = Follow.objects.filter(user=request.user, author=author)
-    if relation.exists():
-        relation.delete()
-    return redirect('profile', username=username)
-
-
-@login_required
 def favorites_view(request, username):
-    favorites = Favorite.objects.filter(user__username=username)
-    recipe_list = Recipe.objects.filter(favorites__in=favorites)
+    tags, tags_filter = get_tags(request)
+    if tags_filter:
+        recipe_list = Recipe.objects.filter(tags_filter).filter(
+            favorites__user__username=username
+        )
+    else:
+        recipe_list = Recipe.objects.filter(favorites__user__username=username)
     paginator = Paginator(recipe_list, 6)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
     return render(
         request,
         'favorites.html',
-        {'page': page, 'paginator': paginator}
+        {'page': page, 'paginator': paginator, 'tags': tags}
     )
 
 
@@ -167,47 +157,7 @@ def purchases_view(request, username):
     return render(
         request,
         'shoplist.html',
-        {'page': page, 'paginator': paginator}
-    )
-
-
-@login_required
-def delete_purchase(request, username, recipe_id):
-    recipe = get_object_or_404(Recipe, id=recipe_id)
-    purchase = Purchase.objects.filter(user=request.user, recipe=recipe)
-    if purchase.exists():
-        purchase.delete()
-    return redirect(
-        'purchases_view',
-        username=username
-    )
-
-
-def filter_tags(request, tag):
-    recipe_list = Recipe.objects.filter(tag__contains=tag)
-    paginator = Paginator(recipe_list, 6)
-    page_number = request.GET.get('page')
-    page = paginator.get_page(page_number)
-    return render(
-        request,
-        'index.html',
-        {'page': page, 'paginator': paginator}
-    )
-
-
-def filter_tags_favorite(request, username, tag):
-    favorites = Favorite.objects.filter(user__username=username)
-    recipe_list = Recipe.objects.filter(
-        favorites__in=favorites,
-        tag__contains=tag
-    )
-    paginator = Paginator(recipe_list, 6)
-    page_number = request.GET.get('page')
-    page = paginator.get_page(page_number)
-    return render(
-        request,
-        'favorites.html',
-        {'page': page, 'paginator': paginator}
+        {'page': page, 'paginator': paginator, 'purchases': purchases}
     )
 
 
@@ -219,7 +169,7 @@ def purchase_list(request):
     ingredient_recipes = IngredientRecipe.objects.filter(
         ingredient__in=ingredients,
         recipe__in=recipes
-        )
+    )
     purchaselist = {}
     for item in ingredient_recipes:
         if purchaselist.get(item.ingredient.name) is None:
@@ -230,9 +180,8 @@ def purchase_list(request):
         else:
             purchaselist[item.ingredient.name][1] += item.value
     response = HttpResponse(content_type='application/pdf')
-    response[
-        'Content-Disposition'
-        ] = 'attachment; filename="purchase_list.pdf"'
+    content = 'attachment; filename="purchase_list.pdf"'
+    response['Content-Disposition'] = content
     p = canvas.Canvas(response)
     for key, value in purchaselist.items():
         p.drawString(100, 100, '{} {} {}'.format(key, value[0], value[1]))
